app/ws_layer/requirements.txt
python-dotenv
websockets
python-socketio
asyncpg
uuid
pydantic
jsonwebtoken
fastapi
pyjwt
requests
websockets
uvicorn